module_2/Modul_2A/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
mysql-connector-python==9.0.0
pika==1.3.2
